Move byte_size_of to new core_common
