+++ lib/vector-core/Cargo.toml
@@ -22,7 +22,7 @@ http = { version = "0.2.5", default-features = false }
 hyper-proxy = { version = "0.9.1", default-features = false, features = ["openssl-tls"] }
 # pinned due to https://github.com/tkaitchuck/aHash/issues/95
 indexmap = { version = "~1.6.2", default-features = false, features = ["serde"] }
-internal_event = { path = "internal-event", default-features = false }
+core_common = { path = "core-common", default-features = false }
 lazy_static = { version = "1.4.0", default-features = false }
 lookup = { path = "../lookup", features = ["arbitrary"] }
 metrics = { version = "0.17.0", default-features = false, features = ["std"]}
