+++ lib/vector-core/core-common/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "internal_event"
+name = "core_common"
 version = "0.1.0"
 authors = ["Vector Contributors <vector@timber.io>"]
 edition = "2018"
