+++ lib/vector-core/core-common/src/byte_size_of.rs
@@ -1,7 +1,4 @@
-use std::{
-    collections::{BTreeMap, BTreeSet},
-    mem,
-};
+use std::{collections::{BTreeMap, BTreeSet}, mem};
 
 pub trait ByteSizeOf {
     /// Returns the in-memory size of this type
