+++ lib/vector-core/core-common/src/internal_event.rs
@@ -1,0 +1,9 @@
+pub trait InternalEvent {
+    fn emit_logs(&self) {}
+    fn emit_metrics(&self) {}
+}
+
+pub fn emit(event: &impl InternalEvent) {
+    event.emit_logs();
+    event.emit_metrics();
+}
