+++ lib/vector-core/core-common/src/lib.rs
@@ -1,0 +1,10 @@
+//! The Vector Core common library
+//!
+//! This library includes common functionality relied upon by vector-core
+//! and core-related crates (e.g. buffers).
+
+#![deny(clippy::all)]
+#![deny(clippy::pedantic)]
+
+pub mod byte_size_of;
+pub mod internal_event;
