+++ lib/vector-core/src/lib.rs
@@ -33,14 +33,13 @@ pub mod source;
 mod test_util;
 pub mod transform;
 pub use buffers;
-mod byte_size_of;
 pub mod partition;
 pub mod serde;
 pub mod stream;
 pub mod time;
-pub use internal_event;
+pub use core_common::internal_event;
+pub use core_common::byte_size_of::ByteSizeOf;
 
-pub use byte_size_of::ByteSizeOf;
 use std::path::PathBuf;
 
 #[macro_use]
